+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- Sync with upstream Substrate ss58 registry
+
+
 ## 8.3.2 Jan 15, 2022
 
 Changes: